requirements.txt
adal==1.2.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
greenlet==1.1.0
idna==2.10
numpy==1.20.3
pandas==1.2.4
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-gitlab==2.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.17
urllib3==1.26.5
